+++ src/components/MainBoard/MainPin.js
