+++ src/components/Modal/Portal.js
@@ -2,10 +2,7 @@ import { useMemo } from "react";
 import { createPortal } from "react-dom";
 
 function Portal({ children, elementId }) {
-    const rootElement = useMemo(
-        () => document.getElementById(elementId),
-        [elementId]
-    );
+    const rootElement = useMemo(() => document.getElementById(elementId), [elementId]);
 
     return createPortal(children, rootElement);
 }
