+++ src/components/index.js
@@ -1,23 +1,28 @@
-import PinImage from "./PinImage";
-import PinDesc from "./PinDesc";
-import PinHeader from "./PinHeader";
-import PinWriterInfo from "./PinWriterInfo";
-import PinSaveInfo from "./PinSaveInfo";
-import PinContainer from "./PinContainer";
-import Comment from "./Comment";
-import CommentEdit from "./CommentEdit";
-import CommentContainer from "./CommentContainer";
-import CommentWrite from "./CommentWrite";
-import CommentList from "./CommentList";
-import DropDown from "./DropDown";
-import Header from "./Header";
-import LoginHeader from "./LoginHeader";
-import LoginMainboard from "./LoginMainboard";
-import Mainboard from "./Mainboard";
-import Modal from "./Modal";
-import Portal from "./Portal";
+import Pin from "./PinDetail/Pin";
+import PinImage from "./PinDetail/PinImage";
+import PinDesc from "./PinDetail/PinDesc";
+import PinHeader from "./PinDetail/PinHeader";
+import PinWriterInfo from "./PinDetail/PinWriterInfo";
+import PinSaveInfo from "./PinDetail/PinSaveInfo";
+import PinContainer from "./PinDetail/PinContainer";
+import Comment from "./Comment/Comment";
+import CommentEdit from "./Comment/CommentEdit";
+import CommentContainer from "./Comment/CommentContainer";
+import CommentWrite from "./Comment/CommentWrite";
+import CommentList from "./Comment/CommentList";
+import LoginForm from "./Login/LoginForm";
+import LoginHeader from "./Login/LoginHeader";
+import LoginMainboard from "./Login/LoginMainboard";
+import DropDown from "./DropDown/DropDown";
+import Header from "./Header/Header";
+import Mainboard from "./MainBoard/Mainboard";
+import MainboardCss from "./MainBoard/Mainboard.css";
+import Modal from "./Modal/Modal";
+import Portal from "./Modal/Portal";
+import SignupCard from "./SignUp/SignupCard";
 
 export {
+    Pin,
     PinImage,
     PinDesc,
     PinHeader,
@@ -31,9 +36,12 @@ export {
     CommentWrite,
     DropDown,
     Header,
+    LoginForm,
     LoginHeader,
     LoginMainboard,
     Mainboard,
+    MainboardCss,
     Modal,
     Portal,
+    SignupCard,
 };
